move newsletter subscription controller to api namespace
