+++ app/views/api/v0/newsletter_subscriptions/create.json.jbuilder
